system: use flatpak for a buncha apps
In the spirit of sandboxing more stuff that's either:

  a) a portal to the internet

  b) non-foss

I've swapped to more Flatpak stuff. This commit includes changes to sway
config to launch flatpaks for the packages it used to use, and the
addition of aliases to every flatpak I have installed so far so they can
be easily launched from terminal if need be. There's also a new `update` alias
that runs both paru and flatpak updates in sequence.